careers handling deletion complete several things todo: - remove unnecessary files from db folder - remove unnecessary files from scripts folder - add back and integrate the controller files in the scripts folder - add actual tests to scripts/tests.js - the small stuff that are some other project objectives such as pagination
