--- routes/errors.js
@@ -1,47 +1,0 @@
-class AppError extends Error {
-    constructor(message, statusCode) {
-        super(message);
-        this.statusCode = statusCode;
-        this.isOperational = true;
-        Error.captureStackTrace(this, this.constructor);
-    }
-}
-
-class NotFoundError extends AppError {
-    constructor(message = 'Resource not found') {
-        super(message, 404);
-    }
-}
-
-class ValidationError extends AppError {
-    constructor(message = 'Validation error') {
-        super(message, 400);
-    }
-}
-
-class UnauthorizedError extends AppError {
-    constructor(message = 'Unauthorized') {
-        super(message, 401);
-    }
-}
-
-class ForbiddenError extends AppError {
-    constructor(message = 'Forbidden') {
-        super(message, 403);
-    }
-}
-
-class ConflictError extends AppError {
-    constructor(message = 'Conflict') {
-        super(message, 409);
-    }
-}
-
-module.exports = {
-    AppError,
-    NotFoundError,
-    ValidationError,
-    UnauthorizedError,
-    ForbiddenError,
-    ConflictError
-};
