+++ scripts/README.md
@@ -1,4 +1,4 @@
 # Scripts
 
 ### About
-The idea of this folder is to have all the scripts, such as parsing scripts and testing scripts and files related to them. There's not much to write on here at the moment because it's just a pool of stuff currently. For future intents and purposes, this can be used as a "MISC" folder
+The idea of this folder is to have all the scripts, such as parsing scripts and testing scripts and utility scripts and files related to them. There's not much to write on here at the moment because it's just a pool of stuff currently. For future intents and purposes, this can be used as a "MISC" or "utils" or "controllers" folder
